Moving maintenance of feeChainSplitter address under params.go
Moved maintenance of fee chain splitter address under a new params.go
file to seperate it from precompile domain.
Added getter to read the current fee chain splitter address.
