+++ precompile/maintenance/maintenance.go
@@ -99,6 +99,7 @@ func newPrecompileMethods(
 
 	// TODO: unclear, do we need to add this function behind EVM/Precompiles flag?
 	methods = append(methods, newSetFeeChainSplitterAddressMethod(poaKeeper, evmKeeper))
+	methods = append(methods, newGetFeeChainSplitterAddressMethod(evmKeeper))
 
 	return methods
 }
